Refactor behaviour system to use BehaviourObject base class
