+++ Assets/Scripts/Tools/Behaviour Tree/Agent.cs
@@ -5,43 +5,21 @@ using RPG;
 
 namespace BehaviourTree
 {
-    public class Agent : MonoBehaviour
+    public class Agent : BehaviourObject
     {
         [SerializeField] private BehaviourTree behaviourTree;
 
-        public Dictionary<string, object> Properties { get; private set; }
-
         private Tree<Behaviour>.Node root;
 
-        void Awake()
+        protected override void Awake()
         {
+            base.Awake();
             root = behaviourTree.Root;
-            Properties = new Dictionary<string, object>();
         }
 
-        void Update()
+        protected void Update()
         {
             root.Element.Tick(root, this);
-        }
-
-        public void SetProperty(string name, object property)
-        {
-            Properties[name] = property;
-        }
-
-        public object GetProperty(string name)
-        {
-            return HasProperty(name) ? Properties[name] : null;
-        }
-
-        public bool HasProperty(string name)
-        {
-            return Properties.ContainsKey(name);
-        }
-
-        public void RemoveProperty(string name)
-        {
-            Properties.Remove(name);
         }
     }
 }
