+++ Assets/Scripts/Tools/Behaviour Tree/BehaviourObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6314624073fd19344a66e0b14acc8eef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
