+++ Assets/Scripts/Tools/Behaviour Tree/Nodes/NotNode.cs
@@ -8,10 +8,10 @@ namespace BehaviourTree
     {
         public void Init(Behaviour behaviour) { }
 
-        public NodeStatus Tick(Tree<Behaviour>.Node self, Agent agent)
+        public NodeStatus Tick(Tree<Behaviour>.Node self, BehaviourObject obj)
         {
             Tree<Behaviour>.Node child = self.GetChild(0);
-            NodeStatus childStatus = child.Element.Tick(child, agent);
+            NodeStatus childStatus = child.Element.Tick(child, obj);
             return childStatus == NodeStatus.Failure ? NodeStatus.Success : NodeStatus.Failure;
         }
     }
