+++ Assets/Scripts/Tools/Behaviour Tree/Nodes/RangeCheckNode.cs
@@ -15,17 +15,17 @@ namespace BehaviourTree
             behaviour.Properties.Add(PROP_POSITION_SRC, new VariableProperty(VariableProperty.Type.String));
         }
 
-        public NodeStatus Tick(Tree<Behaviour>.Node self, Agent agent)
+        public NodeStatus Tick(Tree<Behaviour>.Node self, BehaviourObject obj)
         {
             Behaviour behaviour = self.Element;
 
             // Get the source position
             string src = behaviour.GetProperty(PROP_POSITION_SRC).GetString();
-            if (agent.HasProperty(src))
+            if (obj.HasProperty(src))
             {
                 // Get the distance between source and this agent
-                Vector2 srcPosition = (Vector2)agent.GetProperty(src);
-                float dist = Vector2.Distance(srcPosition, agent.transform.position);
+                Vector2 srcPosition = (Vector2)obj.GetProperty(src);
+                float dist = Vector2.Distance(srcPosition, obj.transform.position);
 
                 // Check whether we're in the specified range
                 float range = (float)behaviour.GetProperty(PROP_RANGE).GetNumber();
